ML/backend/requirements.txt
absl-py==2.1.0
accelerate==1.2.1
asgiref==3.8.1
astunparse==1.6.3
attrs==22.2.0
biopython==1.85
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
discrete-key-value-bottleneck-pytorch==0.1.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
einops==0.8.0
einx==0.3.0
enformer-pytorch==0.8.10
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
frozendict==2.4.6
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
kagglehub==0.3.6
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpi4py==4.0.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.26.0
nbformat==5.10.4
networkx==3.4.2
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
polars==1.17.1
protobuf==5.29.2
psutil==6.1.1
pyfaidx==0.8.1.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vector-quantize-pytorch==1.20.11
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zipp==3.21.0
